django_app/requirements.txt
# Django Core
Django==4.2.7
django-environ==0.11.2

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Celery & Redis
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-redis==5.4.0

# Authentication & Security
django-otp==1.3.0
qrcode[pil]==7.4.2
django-cors-headers==4.3.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Development
ipython==8.18.1
django-extensions==3.2.3



# Channels for WebSocket
channels==4.0.0
daphne==4.0.0
channels-redis==4.1.0

# ROS Bridge
roslibpy==1.5.0
websocket-client==1.6.4

